Make byte handling flexible Py2 => Py3
Postel: "be conservative in what you do, be liberal in what you
accept". We try and follow this principle here. The effect of
ensuring data types are followed strictly moving from Py2 => Py3
seems to be that the margin for error gets even tighter. Instead,
where possible, if we try and convert compatible data types, the
library can be made much more drop-in than it is at present.